+++ src/DpressServices.php
@@ -102,6 +102,7 @@ use Dynart\Dpress\Service\SettingService;
 use Dynart\Dpress\Service\TaxonomyService;
 use Dynart\Dpress\Theme\Places;
 use Dynart\Dpress\Theme\ThemeService;
+use Dynart\Dpress\Theme\ThemeAssets;
 use Dynart\Dpress\Service\RoleService;
 use Dynart\Dpress\Service\SchemaService;
 use Dynart\Dpress\Service\UserService;
@@ -220,6 +221,7 @@ class DpressServices {
         Micro::add(TaxonomyService::class);
         Micro::add(SettingService::class);
         Micro::add(ThemeService::class);
+        Micro::add(ThemeAssets::class);
         Micro::add(PluginService::class);
         Micro::add(MenuService::class);
         Micro::add(Blocks::class);